backend/requirements.txt
asgiref
Django>=4.0,<5.0
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary>=2.9
python-dotenv
# Add other dependencies here